issue 1034 : refactoring
