+++ src/api/mvc/variants/main.go
@@ -1,4 +1,4 @@
-package tables
+package variants
 
 import (
 	"fmt"
@@ -23,6 +23,7 @@ import (
 	"api/models/ingest"
 	"api/mvc"
 	esRepo "api/repositories/elasticsearch"
+	variantService "api/services/variants"
 	"api/utils"
 
 	"api/models/constants/zygosity"
@@ -458,7 +459,11 @@ func VariantsIngest(c echo.Context) error {
 func GetVariantsOverview(c echo.Context) error {
 	fmt.Printf("[%s] - GetVariantsOverview hit!\n", time.Now())
 
-	resultsMap := ObtainVariantsOverview(c)
+	es := c.(*contexts.GohanContext).Es7Client
+	cfg := c.(*contexts.GohanContext).Config
+
+	// TODO: refactor to handle errors better
+	resultsMap := variantService.GetVariantsOverview(es, cfg)
 
 	return c.JSON(http.StatusOK, resultsMap)
 }
@@ -787,77 +792,4 @@ func executeCountByIds(c echo.Context, ids []string, isVariantIdQuery bool) erro
 	}
 
 	return c.JSON(http.StatusOK, respDTO)
-}
-
-// -- helper functions
-func ObtainVariantsOverview(_c echo.Context) map[string]interface{} {
-	resultsMap := map[string]interface{}{}
-	resultsMux := sync.RWMutex{}
-
-	var wg sync.WaitGroup
-	es := _c.(*contexts.GohanContext).Es7Client
-	cfg := _c.(*contexts.GohanContext).Config
-
-	callGetBucketsByKeyword := func(key string, keyword string, _wg *sync.WaitGroup) {
-		defer _wg.Done()
-
-		results, bucketsError := esRepo.GetVariantsBucketsByKeywordAndTableId(cfg, es, keyword, "")
-		if bucketsError != nil {
-			resultsMap[key] = map[string]interface{}{
-				"error": "Something went wrong. Please contact the administrator!",
-			}
-			return
-		}
-
-		// retrieve aggregations.items.buckets
-		bucketsMapped := []interface{}{}
-		if aggs, aggsOk := results["aggregations"]; aggsOk {
-			aggsMapped := aggs.(map[string]interface{})
-
-			if items, itemsOk := aggsMapped["items"]; itemsOk {
-				itemsMapped := items.(map[string]interface{})
-
-				if buckets, bucketsOk := itemsMapped["buckets"]; bucketsOk {
-					bucketsMapped = buckets.([]interface{})
-				}
-			}
-		}
-
-		individualKeyMap := map[string]interface{}{}
-		// push results bucket to slice
-		for _, bucket := range bucketsMapped {
-			doc_key := fmt.Sprint(bucket.(map[string]interface{})["key"]) // ensure strings and numbers are expressed as strings
-			doc_count := bucket.(map[string]interface{})["doc_count"]
-
-			individualKeyMap[doc_key] = doc_count
-		}
-
-		resultsMux.Lock()
-		resultsMap[key] = individualKeyMap
-		resultsMux.Unlock()
-	}
-
-	// get distribution of chromosomes
-	wg.Add(1)
-	go callGetBucketsByKeyword("chromosomes", "chrom.keyword", &wg)
-
-	// get distribution of variant IDs
-	wg.Add(1)
-	go callGetBucketsByKeyword("variantIDs", "id.keyword", &wg)
-
-	// get distribution of sample IDs
-	wg.Add(1)
-	go callGetBucketsByKeyword("sampleIDs", "sample.id.keyword", &wg)
-
-	// get distribution of assembly IDs
-	wg.Add(1)
-	go callGetBucketsByKeyword("assemblyIDs", "assemblyId.keyword", &wg)
-
-	// get distribution of table IDs
-	wg.Add(1)
-	go callGetBucketsByKeyword("tableIDs", "tableId.keyword", &wg)
-
-	wg.Wait()
-
-	return resultsMap
 }
