+++ src/api/repositories/elasticsearch/tables.go
@@ -1,7 +1,8 @@
 package elasticsearch
 
 import (
-	"api/contexts"
+	// "api/contexts"
+	"api/models"
 	"api/models/dtos"
 	"api/models/indexes"
 	"api/models/schemas"
@@ -18,17 +19,17 @@ import (
 	"strings"
 	"time"
 
+	es7 "github.com/elastic/go-elasticsearch/v7"
+
 	"github.com/elastic/go-elasticsearch/esapi"
 	"github.com/google/uuid"
-	"github.com/labstack/echo"
 	"github.com/mitchellh/mapstructure"
 )
 
 const tablesIndex = "tables"
 
-func CreateTable(c echo.Context, t dtos.CreateTableRequestDto) (indexes.Table, error) {
+func CreateTable(es *es7.Client, ctxt context.Context, t dtos.CreateTableRequestDto) (indexes.Table, error) {
 
-	es := c.(*contexts.GohanContext).Es7Client
 	now := time.Now()
 
 	// TODO: improve checks and balances..
@@ -77,7 +78,7 @@ func CreateTable(c echo.Context, t dtos.CreateTableRequestDto) (indexes.Table, e
 	fmt.Println(reflect.TypeOf(req))
 
 	// Return an API response object from request
-	res, err := req.Do(c.Request().Context(), es)
+	res, err := req.Do(ctxt, es)
 	if err != nil {
 		fmt.Printf("IndexRequest ERROR: %s\n", err)
 		return docStruct, err
@@ -108,10 +109,7 @@ func CreateTable(c echo.Context, t dtos.CreateTableRequestDto) (indexes.Table, e
 	return docStruct, nil
 }
 
-func GetTables(c echo.Context, tableId string, dataType string) (map[string]interface{}, error) {
-
-	cfg := c.(*contexts.GohanContext).Config
-	es := c.(*contexts.GohanContext).Es7Client
+func GetTables(cfg *models.Config, es *es7.Client, ctxt context.Context, tableId string, dataType string) (map[string]interface{}, error) {
 
 	if cfg.Debug {
 		http.DefaultTransport.(*http.Transport).TLSClientConfig = &tls.Config{InsecureSkipVerify: true}
@@ -194,10 +192,7 @@ func GetTables(c echo.Context, tableId string, dataType string) (map[string]inte
 
 }
 
-func GetTablesByName(c echo.Context, tableName string) ([]indexes.Table, error) {
-
-	cfg := c.(*contexts.GohanContext).Config
-	es := c.(*contexts.GohanContext).Es7Client
+func GetTablesByName(cfg *models.Config, es *es7.Client, ctxt context.Context, tableName string) ([]indexes.Table, error) {
 
 	allTables := make([]indexes.Table, 0)
 
@@ -300,10 +295,7 @@ func GetTablesByName(c echo.Context, tableName string) ([]indexes.Table, error) 
 	return allTables, nil
 }
 
-func DeleteTableById(c echo.Context, tableId string) (map[string]interface{}, error) {
-
-	cfg := c.(*contexts.GohanContext).Config
-	es := c.(*contexts.GohanContext).Es7Client
+func DeleteTableById(cfg *models.Config, es *es7.Client, ctxt context.Context, tableId string) (map[string]interface{}, error) {
 
 	if cfg.Debug {
 		http.DefaultTransport.(*http.Transport).TLSClientConfig = &tls.Config{InsecureSkipVerify: true}
